Move some coin's RPCs to separate modules * Move `account_balance`, `init_withdraw`, `init_create_account` RPCs from `coins` to `coins::rpc_command` * Use `RpcTask` for `init_scan_for_new_addresses`
