+++ mm2src/coins/rpc_command/init_create_account.rs
