+++ mm2src/coins/rpc_command/init_withdraw.rs
