+++ mm2src/rpc/dispatcher/dispatcher.rs
@@ -6,12 +6,13 @@ use crate::{mm2::lp_stats::{add_node_to_version_stat, remove_node_from_version_s
                             stop_version_stat_collection, update_version_stat_collection},
             mm2::lp_swap::{recreate_swap_data, trade_preimage_rpc},
             mm2::rpc::get_public_key::get_public_key};
-use coins::coin_balance::{account_balance, scan_for_new_addresses};
 use coins::hd_wallet::get_new_address;
-use coins::init_create_account::{init_create_new_account, init_create_new_account_status,
-                                 init_create_new_account_user_action};
-use coins::init_withdraw::{init_withdraw, withdraw_status, withdraw_user_action};
 use coins::my_tx_history_v2::my_tx_history_v2_rpc;
+use coins::rpc_command::account_balance::account_balance;
+use coins::rpc_command::init_create_account::{init_create_new_account, init_create_new_account_status,
+                                              init_create_new_account_user_action};
+use coins::rpc_command::init_scan_for_new_addresses::{init_scan_for_new_addresses, init_scan_for_new_addresses_status};
+use coins::rpc_command::init_withdraw::{init_withdraw, withdraw_status, withdraw_user_action};
 use coins::utxo::bch::BchCoin;
 use coins::utxo::qtum::QtumCoin;
 use coins::utxo::slp::SlpToken;
@@ -143,7 +144,8 @@ async fn dispatcher_v2(request: MmRpcRequest, ctx: MmArc) -> DispatcherResult<Re
         "recreate_swap_data" => handle_mmrpc(ctx, request, recreate_swap_data).await,
         "remove_delegation" => handle_mmrpc(ctx, request, remove_delegation).await,
         "remove_node_from_version_stat" => handle_mmrpc(ctx, request, remove_node_from_version_stat).await,
-        "scan_for_new_addresses" => handle_mmrpc(ctx, request, scan_for_new_addresses).await,
+        "init_scan_for_new_addresses" => handle_mmrpc(ctx, request, init_scan_for_new_addresses).await,
+        "init_scan_for_new_addresses_status" => handle_mmrpc(ctx, request, init_scan_for_new_addresses_status).await,
         "start_simple_market_maker_bot" => handle_mmrpc(ctx, request, start_simple_market_maker_bot).await,
         "start_version_stat_collection" => handle_mmrpc(ctx, request, start_version_stat_collection).await,
         "stop_simple_market_maker_bot" => handle_mmrpc(ctx, request, stop_simple_market_maker_bot).await,
